move scrum classes into folder
