+++ lib/scrum.rb
@@ -1,0 +1,10 @@
+module Scrum
+end
+
+require 'scrum/card_type_detection'
+require 'scrum/sprint_board'
+require 'scrum/sprint_planning_board'
+
+require 'scrum/prioritizer'
+require 'scrum/sprint_cleanup'
+require 'scrum/move_backlog'
