+++ lib/trollolo.rb
@@ -33,13 +33,7 @@ require_relative 'burndown_data'
 require_relative 'backup'
 require_relative 'checklist'
 require_relative 'trello_service'
-require_relative 'prioritizer'
-require_relative 'card_type_detection'
-require_relative 'sprint_board'
-require_relative 'sprint_planning_board'
-require_relative 'move_backlog'
-require_relative 'sprint_cleanup'
-require_relative 'move_backlog'
+require_relative 'scrum'
 
 class TrolloloError < StandardError
 end
